requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=3.0
Flask-Login>=0.5
python-dotenv>=0.19
requests>=2.25
psycopg2-binary>=2.9
Werkzeug>=2.3.7
Pillow>=10.1.0
python-dateutil>=2.8.2
Markdown>=3.3.6
gunicorn>=21.2.0
SQLAlchemy>=2.0.23
itsdangerous>=2.1.2
Jinja2>=3.1.2
alembic>=1.12.1
pytz>=2023.3
xhtml2pdf>=0.2.11
bcrypt>=4.1.3
Flask-WTF>=1.0
email_validator>=2.0
Flask-Limiter>=3.6.0
Flask-Bcrypt>=1.0
reportlab>=4.0
weasyprint>=60
stripe>=7.0,<8.0